requirements/base.txt
# Core
Django>=2.0,<2.1
wagtail>=2.2,<2.3

# Database
# NOTE: version 2.7.4 is busted.. do not update.
psycopg2==2.7.5

django-model-utils==3.1.2

# Reading environment variables, secure enclaves
django-environ==0.4.5
#django-secure==1.0.1

# Static and Media files
django-storages==1.7
boto3==1.8.6
django-pipeline==1.6.14

# Utilities
pytz==2018.5
Pillow==5.2.0

django-suit==0.2.26

django-webpack-loader==0.6.0